chore: remove WS implementation (for comparison purposes)
