+++ cmd/server/main.go
@@ -1,24 +1,71 @@
 package main
 
 import (
+	"context"
+	"encoding/json"
 	"fmt"
+	"io"
 	"log"
 	"net/http"
+	"time"
 
-	"github.com/benchttp/engine/server"
+	"github.com/benchttp/engine/internal/configparse"
+	"github.com/benchttp/engine/runner"
 )
 
 const (
 	port = "8080"
-	// token is a dummy token used for development only.
-	token = "6db67fafc4f5bf965a5a" //nolint:gosec
 )
 
 func main() {
 	addr := ":" + port
 	fmt.Println("http://localhost" + addr)
 
-	handler := server.NewHandler(false, token)
+	log.Fatal(http.ListenAndServe(addr, http.HandlerFunc(handleStream)))
+}
 
-	log.Fatal(http.ListenAndServe(addr, handler))
+func handleStream(w http.ResponseWriter, r *http.Request) {
+	w.Header().Set("Access-Control-Allow-Origin", "*")
+
+	b, err := io.ReadAll(r.Body)
+	if err != nil {
+		internalError(w, err)
+		return
+	}
+
+	cfg, err := configparse.JSON(b)
+	if err != nil {
+		internalError(w, err)
+		return
+	}
+
+	rep, err := runner.New(streamProgress(w)).Run(context.Background(), cfg)
+	if err != nil {
+		internalError(w, err)
+		return
+	}
+
+	// hack: the write for the Report sometimes appears to be merged
+	// with the last write for the Progress, leading to invalid JSON.
+	// The issue is likely on the read side (front-end), but this is
+	// the easiest fix for now.
+	time.Sleep(10 * time.Millisecond)
+	if err := json.NewEncoder(w).Encode(rep); err != nil {
+		internalError(w, err)
+		return
+	}
+}
+
+func streamProgress(w http.ResponseWriter) func(runner.RecordingProgress) {
+	enc := json.NewEncoder(w)
+	return func(progress runner.RecordingProgress) {
+		if err := enc.Encode(progress); err != nil {
+			internalError(w, err)
+		}
+		w.(http.Flusher).Flush()
+	}
+}
+
+func internalError(w http.ResponseWriter, err error) {
+	http.Error(w, err.Error(), http.StatusInternalServerError)
 }
