+++ go.mod
@@ -8,7 +8,4 @@ require (
 	gopkg.in/yaml.v3 v3.0.1
 )
 
-require (
-	github.com/drykit-go/cond v0.1.0 // indirect
-	github.com/gorilla/websocket v1.5.0
-)
+require github.com/drykit-go/cond v0.1.0 // indirect
